requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
notion-client
pydantic
google-generativeai  # Gemini を使うなら
